requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-yaml==2.0.0
drf-spectacular==0.25.1
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
